agents/requirements.txt
solace-pubsub-plus==10.21.0
python-dotenv==1.0.0
requests==2.31.0
pytrends==4.9.2
praw==7.7.1
python-dateutil==2.8.2
pysolace==1.0.0
typing-extensions==4.7.1
